User authentization

+++ dao/src/main/java/cz/muni/fi/pa165/music_library/dao/implementations/UserDaoImpl.java
@@ -56,6 +56,7 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     public void deleteUser(User user) {
+        if (user == null) throw new IllegalArgumentException();
         em.remove(em.find(User.class, user.getUserId()));
     }
